Fix commands (need to refactor)

+++ dnevnichok/commands.py
@@ -1,3 +1,4 @@
+import curses
 import os
 import sys
 import sqlite3
@@ -5,6 +6,7 @@ import inspect
 import logging
 
 from dnevnichok.config import Config
+from dnevnichok.core import NoteItem, TagItem
 from dnevnichok.events import event_hub
 from dnevnichok.populate import parse_note
 
@@ -35,34 +37,62 @@ class deleteCommand(Command):
         self.executor = executor
         self.conn = sqlite3.connect(dbpath)
 
-        self.item = None
+        self.item = self.executor.app.window.get_current_item()
 
     def ensure(self):
         return self.executor.ensure('Are you sure you want to delete {}? [y/N]  '.format(self.item.get_path()))
 
     def run(self):
-        exit_code = os.system('rm ' + self.item.get_path())
-        if exit_code == 0:
+        table = None
+        if isinstance(self.item, NoteItem):
+            exit_code = os.system('rm ' + self.item.get_path())
+            if exit_code == 0:
+                table = 'notes'
+        elif isinstance(self.item, TagItem):
+            table = 'tags'
+
+        if table:
             with self.conn:
                 cur = self.conn.cursor()
-                cur.execute('DELETE FROM notes WHERE id = {}'.format(self.item.id))
-                event_hub.trigger(('reload',))
+                cur.execute('DELETE FROM {} WHERE id = {}'.format(table, self.item.id))
+            event_hub.trigger(('reload',))
+            curses.curs_set(1)  # THIS is sought-for hack
+            curses.curs_set(0)
+
+
+class commitCommand(Command):
+    def __init__(self, executor, args: tuple):
+        self.executor = executor
+
+    def run(self):
+        os.system('git add .')
+        os.system('git commit')
+        event_hub.trigger(('reload',))
+        curses.curs_set(1)  # THIS is sought-for hack
+        curses.curs_set(0)
 
 
 class newCommand(Command):
     def __init__(self, executor, args: tuple):
         self.executor = executor
         self.conn = sqlite3.connect(dbpath)
+        self.path = None
 
-        try:
-            self.item_title = args[0]
-        except IndexError:
+        args_num = len(args)
+        if args_num == 0:
             raise InsufficientArguments(1)
+        if args_num > 0:
+            self.item_title = args[0] if args[0].endswith('.rst') else args[0] + '.rst'
+        if args_num > 1:
+            self.path = args[1]
 
     def run(self):
-        dir_id = self.executor.app.file_manager.base
-        base_path = self.executor.app.file_manager.get_current_path()
-        note_path = os.path.join(base_path, self.item_title)
+        base_path = self.executor.app.manager_hub.get_current_dir() if not self.path else self.path
+        dir_id = self.executor.app.manager_hub.get_path_id(base_path)
+        note_path = os.path.join('.', base_path, self.item_title)
+        if os.path.exists(note_path):
+            event_hub.trigger(('print', 'File {} already exists'.format(note_path)))
+            return
         os.system('touch ' + note_path)
         exit_code = os.system('vi ' + note_path)
         if exit_code == 0:
@@ -76,7 +106,10 @@ class newCommand(Command):
                 for tag in note.tags:
                     cur.execute("INSERT OR IGNORE INTO tags(title) VALUES(?)", (tag,))
                     cur.execute("INSERT INTO note_tags(note_id, tag_id) VALUES(?, ?)", (note.id, cur.lastrowid,))
-                event_hub.trigger(('reload',))
+
+            event_hub.trigger(('reload',))
+            curses.curs_set(1)  # THIS is sought-for hack
+            curses.curs_set(0)
 
 
 def get_all_commands() -> dict:
@@ -89,3 +122,48 @@ def get_all_commands() -> dict:
             if len(name) > 7 and name.endswith('Command'):
                 commands.update({name[:-7]: obj})
     return commands
+
+
+class Executor:
+    """
+    Responsible for take find all available commands, pick right one and
+    give it everything it needs. Also stores history.
+    """
+    def __init__(self, app):
+        self.app = app
+        self.commands = get_all_commands()
+
+    def run_command(self, command: str):
+        command, *args = command.split()
+        try:
+            CommandClass = self.commands[command]
+        except KeyError:
+            event_hub.trigger(('print', "Unknown command: " + command))
+            return
+
+        try:
+            execution = CommandClass(self, args)
+        except InsufficientArguments as e:
+            event_hub.trigger(('print', e.message))
+            return
+
+        proceed = execution.ensure()
+        if proceed:
+            execution.run()
+
+    def ensure(self, text, tries=0):
+        choice = self.app.window.input(text)
+        if choice in 'yY':
+            return True
+        elif choice in 'nN':
+            return False
+        elif tries > 3:
+            event_hub.trigger(('print', "Ok. Just have a good day"))
+            return False
+        else:
+            return self.ensure(text, tries+1)
+
+    def get_current_item(self):
+        return self.app.window.get_current_item()
+
+
